refactor: testing (#14)
* refactor: disabled some html validation aspects

either because of example related or specific usage

* refactor: we don't need to declare this role

as it's set implicitly, compare to e.g. developer.mozilla.org/en-US/docs/Web/HTML/Element/footer

* refactor: we don't need to declare this role

as it's set implicitly, compare to e.g. developer.mozilla.org/en-US/docs/Web/HTML/Element/main

* refactor: disabled the correct code line

* refactor: disabled html validation aspect

* refactor: consistent quotes

* refactor: consistent quotes

* refactor: in case we don't have a description, we won't need the aria-describedby attribute

* chore: providing another default

* refactor: set a default for the buttontype

* refactor: restructured the sample page

to reflect a correct headline structure

* refactor: ignore those headline structures on HTML validation

as they are too hard to influence in sample pages

* refactor: we don't need to declare this role

as it's set implicitly, compare to e.g. developer.mozilla.org/en-US/docs/Web/HTML/Element/nav

* refactor: ignore those headline structures on HTML validation

as they are too hard to influence in sample pages

* refactor: ignore this headline structures on HTML validation

as a headline of type 2 is perfectly fine in this new context

* refactor: there shouldn't be more than one main landmark elements

* refactor: ignore those headline structures on HTML validation

as they are too hard to influence in sample pages

* refactor: details tag nesting in p not allowed

Element <details> is not permitted as content in <p>

* refactor: corrected that attributes naming

* refactor: ignore this headline structures on HTML validation

as they are too hard to influence in sample pages

* refactor: adapted the HTML validation rules to our needs

* refactor: optimized the test and lint scripts

reordered, renamed and added npx

* chore: removed leftover

* refactor: optimized list of test urls for accessibility

* refactor: optimized test scripts

and added new dependency for running server and tests in parallel

* refactor: disabling stylelint rules for specific code parts

* refactor: stylelint feedback

* refactor: stylelint feedback

* refactor: html validate feedback

* test: we don't want to test for self-closing tags

* feat: suggesting html validate as well as VS Code extension

* refactor(test): using the smaller markup-only files for HTML testing

* chore: regenerated package-lock.json

* chore: prettier optimization

* refactor: running prettier for linting

* chore: we need the SCSS function here

* chore: we need to provide different ID values here